Builders, zipper, and phone-gap
Sharpened builders, and created multi_build, for abstraction of the
multiple builders (html and latex) to one builder.
created zipper, which decorates the html builder, meaning it does
exactly the same, only it zips the milon at the end.
created PhoneGap builder, which decorates the zipper, meaning it creates
the .apk at the end.

Also, now with the decorators, it is easy to decide which actions we
want to perform in the pipeline. Just comment out the parts that are
responsible for those actions.